Outsource recipe data
