+++ meal-planner/src/components/recipeData.js
@@ -1,0 +1,15 @@
+import { v4 as uuidv4 } from 'uuid'
+
+const meals = [
+    { meal: 'Käsestulle', id: uuidv4() },
+    { meal: 'Kartoffelsalat', id: uuidv4() },
+    { meal: 'Gemüselasagne', id: uuidv4() },
+    { meal: 'Steinpilzrisotto', id: uuidv4() },
+    { meal: 'Currywurst', id: uuidv4() },
+    { meal: 'Linsensuppe', id: uuidv4() },
+    { meal: 'Spaghetti Carbonara', id: uuidv4() },
+    { meal: 'Banana Pancakes', id: uuidv4() },
+    { meal: 'Thai Gemüsecurry', id: uuidv4() },
+]
+
+export default meals
